Rename project directory to scripts
